requirements.txt
wheel
Flask
flask_sqlalchemy
email-validator==2.0.0.post2
argon2-cffi==20.1.0
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Pillow==9.5.0
python-dotenv==1.0.0
regex==2023.5.5
requests==2.30.0
SQLAlchemy
typing_extensions==4.10.0
urllib3==2.0.2
Werkzeug==2.3.4
WTForms
flask_migrate
flask_caching
authlib
Flask-APScheduler
wtforms-sqlalchemy
schedule
psycopg2-binary
greenlet==3.0.3
typing_extensions==4.10.0
sqlalchemy_utils
azure-identity
azure-mgmt-compute